+++ modules/fork-me-github/CHANGELOG.md
@@ -1,5 +1,16 @@
 # CHANGELOG
 
+## 1.0.1 - 2017-03-17
+
+* Added `config/hooks.php` file.
+* Added `config/routes.php` file.
+
+* Deleted `App\Modules\ForkMeGitHub\Hooks` class.
+* Deleted `App\Modules\ForkMeGitHub\Hooks::add()` method.
+
+* Deleted `App\Modules\ForkMeGitHub\Routes` class.
+* Deleted `App\Modules\ForkMeGitHub\Routes::add()` method.
+
 ## 1.0.0 - 2017-03-09
 
 * Added `App\Modules\ForkMeGitHub\Controller\ForkMeGitHub` class.